# backend/openapi-spec.yml
openapi: 3.0.3
info:
  title: Kanto-Feo Academy API
  description: API pour la gestion de l'académie de musique Kanto-Feo
  version: 1.0.0
  contact:
    name: Kanto-Feo Team
    email: support@kanto-feo.com
servers:
  - url: http://localhost:3000
    description: Serveur de développement
  - url: https://api.kanto-feo.com
    description: Serveur de production
paths:
...
  /admin:
    get:
      tags:
        - Admin
      summary: Récupérer tous les administrateurs
      description: Retourne la liste de tous les administrateurs avec leurs étudiants et cours
      responses:
        '200':
          description: Liste des administrateurs récupérée avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Admin'
    post:
      tags:
        - Admin
      summary: Créer un nouvel administrateur
      description: Crée un nouvel administrateur avec hashage du mot de passe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdminDto'
      responses:
        '201':
          description: Administrateur créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Admin'
        '400':
          description: Données invalides
        '409':
          description: Email déjà utilisé
  /admin/{id}:
    get:
      tags:
        - Admin
      summary: Récupérer un administrateur par ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Administrateur trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Admin'
        '404':
          description: Administrateur non trouvé
    patch:
      tags:
        - Admin
      summary: Mettre à jour un administrateur
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAdminDto'
      responses:
        '200':
          description: Administrateur mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Admin'
        '404':
          description: Administrateur non trouvé
    delete:
      tags:
        - Admin
      summary: Supprimer un administrateur
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Administrateur supprimé
        '404':
          description: Administrateur non trouvé
  /admin/email/{email}:
    get:
      tags:
        - Admin
      summary: Récupérer un administrateur par email
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Administrateur trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Admin'
        '404':
          description: Administrateur non trouvé
  /student:
    get:
      tags:
        - Student
      summary: Récupérer tous les étudiants
      description: Retourne la liste de tous les étudiants avec leurs relations
      responses:
        '200':
          description: Liste des étudiants récupérée avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
    post:
      tags:
        - Student
      summary: Créer un nouvel étudiant
      description: Crée un nouvel étudiant avec hashage du mot de passe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudentDto'
      responses:
        '201':
          description: Étudiant créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          description: Données invalides
        '409':
          description: Email déjà utilisé
  /student/{id}:
    get:
      tags:
        - Student
      summary: Récupérer un étudiant par ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Étudiant trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '404':
          description: Étudiant non trouvé
    patch:
      tags:
        - Student
      summary: Mettre à jour un étudiant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStudentDto'
      responses:
        '200':
          description: Étudiant mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '404':
          description: Étudiant non trouvé
    delete:
      tags:
        - Student
      summary: Supprimer un étudiant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Étudiant supprimé
        '404':
          description: Étudiant non trouvé
  /student/email/{email}:
    get:
      tags:
        - Student
      summary: Récupérer un étudiant par email
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Étudiant trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '404':
          description: Étudiant non trouvé
...
  /payment/student/{studentId}:
    get:
      tags:
        - Payment
      summary: Récupérer les paiements d'un étudiant
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Liste des paiements de l'étudiant
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'
...
  /enrollment/student/{studentId}:
    get:
      tags:
        - Enrollment
      summary: Récupérer les inscriptions d'un étudiant
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Liste des inscriptions de l'étudiant
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Enrollment'
...
  /schedule/student/{studentId}:
    get:
      tags:
        - Schedule
      summary: Récupérer les emplois du temps d'un étudiant
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Liste des emplois du temps de l'étudiant
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schedule'
...
  /progress/student/{studentId}:
    get:
      tags:
        - Progress
      summary: Récupérer les progrès d'un étudiant
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Liste des progrès de l'étudiant
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Progress'
...
  /notification/student/{studentId}:
    get:
      tags:
        - Notification
      summary: Récupérer les notifications d'un étudiant
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Liste des notifications de l'étudiant
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
...
components:
  schemas:
    Admin:
      type: object
      properties:
        admin_id:
          type: integer
          readOnly: true
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        registration_date:
          type: string
          format: date-time
          readOnly: true
        students:
          type: array
          items:
            $ref: '#/components/schemas/Student'
        courses:
          type: array
          items:
            $ref: '#/components/schemas/Course'
    CreateAdminDto:
      type: object
      required:
        - email
        - password
      properties:
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        password:
          type: string
          format: password
    UpdateAdminDto:
      type: object
      properties:
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        password:
          type: string
          format: password
    Student:
      type: object
      properties:
        student_id:
          type: integer
          readOnly: true
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        registration_date:
          type: string
          format: date-time
          readOnly: true
        current_level:
          $ref: '#/components/schemas/CourseLevel'
          nullable: true
        instrument:
          $ref: '#/components/schemas/Instrument'
        admin:
          $ref: '#/components/schemas/Admin'
        enrollments:
          type: array
          items:
            $ref: '#/components/schemas/Enrollment'
        payments:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
        schedules:
          type: array
          items:
            $ref: '#/components/schemas/Schedule'
        progresses:
          type: array
          items:
            $ref: '#/components/schemas/Progress'
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
    CreateStudentDto:
      type: object
      required:
        - email
        - password
        - instrument_id
        - admin_id
      properties:
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        password:
          type: string
          format: password
        current_level:
          $ref: '#/components/schemas/CourseLevel'
          nullable: true
        instrument_id:
          type: integer
        admin_id:
          type: integer
    UpdateStudentDto:
      type: object
      properties:
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        password:
          type: string
          format: password
        current_level:
          $ref: '#/components/schemas/CourseLevel'
          nullable: true
        instrument_id:
          type: integer
        admin_id:
          type: integer
    Instrument:
      type: object
      properties:
        instrument_id:
          type: integer
          readOnly: true
        name:
          type: string
        description:
          type: string
          nullable: true
        students:
          type: array
          items:
            $ref: '#/components/schemas/Student'
        courses:
          type: array
          items:
            $ref: '#/components/schemas/Course'
...
    Course:
      type: object
      properties:
        course_id:
          type: integer
          readOnly: true
        title:
          type: string
        description:
          type: string
          nullable: true
        level:
          $ref: '#/components/schemas/CourseLevel'
          nullable: true
        video_url:
          type: string
          nullable: true
        publication_date:
          type: string
          format: date-time
          readOnly: true
        instrument:
          $ref: '#/components/schemas/Instrument'
        admin:
          $ref: '#/components/schemas/Admin'
        enrollments:
          type: array
          items:
            $ref: '#/components/schemas/Enrollment'
        progresses:
          type: array
          items:
            $ref: '#/components/schemas/Progress'
...
    Payment:
      type: object
      properties:
        payment_id:
          type: integer
          readOnly: true
        amount:
          type: number
          format: float
        payment_date:
          type: string
          format: date-time
          readOnly: true
        status:
          $ref: '#/components/schemas/PaymentStatus'
          nullable: true
        method:
          $ref: '#/components/schemas/PaymentMethod'
          nullable: true
        student:
          $ref: '#/components/schemas/Student'
...
    Enrollment:
      type: object
      properties:
        enrollment_id:
          type: integer
          readOnly: true
        enrollment_date:
          type: string
          format: date-time
          readOnly: true
        student:
          $ref: '#/components/schemas/Student'
        course:
          $ref: '#/components/schemas/Course'
...
    Schedule:
      type: object
      properties:
        schedule_id:
          type: integer
          readOnly: true
        day:
          $ref: '#/components/schemas/DayOfWeek'
          nullable: true
        start_time:
          type: string
          format: date-time
          nullable: true
        end_time:
          type: string
          format: date-time
          nullable: true
        student:
          $ref: '#/components/schemas/Student'
...
    Progress:
      type: object
      properties:
        progress_id:
          type: integer
          readOnly: true
        lesson_completed:
          type: boolean
          default: false
        validation_date:
          type: string
          format: date-time
          nullable: true
        achieved_level:
          $ref: '#/components/schemas/CourseLevel'
          nullable: true
        student:
          $ref: '#/components/schemas/Student'
        course:
          $ref: '#/components/schemas/Course'
...
    Notification:
      type: object
      properties:
        notification_id:
          type: integer
          readOnly: true
        message:
          type: string
          nullable: true
        sent_date:
          type: string
          format: date-time
          readOnly: true
        type:
          $ref: '#/components/schemas/NotificationType'
          nullable: true
        student:
          $ref: '#/components/schemas/Student'
...
    CourseLevel:
      type: string
      enum:
        - BEGINNER
        - INTERMEDIATE
        - ADVANCED
    PaymentStatus:
      type: string
      enum:
        - VALIDATED
        - PENDING
        - REJECTED
    PaymentMethod:
      type: string
      enum:
        - ORANGE_MONEY
        - AIRTEL_MONEY
        - MVOLA
    NotificationType:
      type: string
      enum:
        - REMINDER
        - CONFIRMATION
        - INFO
    DayOfWeek:
      type: string
      enum:
        - MONDAY
        - TUESDAY
        - WEDNESDAY
        - THURSDAY
        - FRIDAY
        - SATURDAY
        - SUNDAY
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...